Refactor method download_from_blobstore
The test deleted in this commit could be deleted, because by using a dedicated http client, the fact that redirects are followed correclty is the responsibility of the http client. Since we are using the standard http client, we can assume that this works. The other test could be deleted, because is too implementation specific. But we don't care about the exact implementation as long as the blob is downloaded into the destination_path.