Quick fix for timeouts between CC and Bits-Service
The default timeout for Ruby's Net::HTTP is 60s, which is not enough to upload things like buildpacks. A better solution is to make this configurable like it is done in the ResourcePool. But this requires a change in CC, which will probably take a bit until it is merged.

[#138382323]